+++ src/PresenterLayer/BatteryPresenter/BatteryPresenter.h
@@ -35,10 +35,10 @@ signals:
     void internalTemperatureReceived(const int&);
     void fanSpeedReceived(const int&);
     void requestedFanSpeedReceived(const int&);
-    void lowCellVoltageReceived(const int&);
+    void lowCellVoltageReceived(const float&);
     void lowCellVoltageIdReceived(const int&);
-    void highCellVoltageReceived(const int&);
+    void highCellVoltageReceived(const float&);
     void highCellVoltageIdReceived(const int&);
-    void averageCellVoltageReceived(const int&);
+    void averageCellVoltageReceived(const float&);
     void packNetPowerReceived(const double&);
 };
